Prevent loosing money by parsing small values (#474)
* Prevent loosing money by parsing small values

* Fix test for explicitly loose fractions

* Add more test for Money.TryParse

+++ NBitcoin/FeeRate.cs
@@ -65,7 +65,7 @@ namespace NBitcoin
 		{
 			if(satoshiPerByte < 0)
 				throw new ArgumentOutOfRangeException("satoshiPerByte");
-			_FeePerK = Money.Satoshis(satoshiPerByte * 1000);
+			_FeePerK = Money.Satoshis((long)(satoshiPerByte * 1000));
 		}
 
 		/// <summary>
